+++ spring-application/build.gradle.kts
@@ -67,6 +67,7 @@ dependencies {
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-reactor")
     implementation("org.jetbrains.kotlinx:kotlinx-serialization-json")
     implementation("org.springframework.boot:spring-boot-starter-webflux")
+    implementation("io.swagger.parser.v3:swagger-parser")
     implementation(projects.springBootTitilerStarterCore)
 
     testImplementation("org.springframework.boot:spring-boot-starter-test") {
