+++ spring-application/src/main/resources/META-INF/spring/aot.factories
@@ -1,0 +1,2 @@
+org.springframework.aot.hint.RuntimeHintsRegistrar=\
+  dev.tronto.titiler.spring.application.swagger.SwaggerRuntimeHints
