+++ spring-boot-titiler-autoconfigure/src/main/resources/META-INF/spring/aot.factories
@@ -1,5 +1,4 @@
 org.springframework.aot.hint.RuntimeHintsRegistrar=\
   dev.tronto.titiler.spring.autoconfigure.hints.KotlinSerializableRuntimeHints,\
   dev.tronto.titiler.spring.autoconfigure.hints.ReactorRuntimeHints,\
-  dev.tronto.titiler.spring.autoconfigure.hints.ValueClassRuntimeHints,\
-  dev.tronto.titiler.spring.autoconfigure.hints.SwaggerRuntimeHints
+  dev.tronto.titiler.spring.autoconfigure.hints.ValueClassRuntimeHints
