+++ spring-boot-titiler-starter-core/build.gradle.kts
@@ -18,7 +18,6 @@ dependencies {
     api("org.thymeleaf:thymeleaf")
     api("org.springframework:spring-webflux")
     api("org.jetbrains.kotlinx:kotlinx-serialization-core")
-    api("io.swagger.parser.v3:swagger-parser")
 }
 
 kotlin {
